requirement.txt
Django==2.0.7
django-cors-headers==2.3.0
djangorestframework==3.8.2
Markdown==2.6.11
mysqlclient==1.3.13
pytz==2018.5
